Fix integer overflow in Gadget block reading for large particle counts
- Cast N*block.dimension to np.int64 to prevent overflow when N > 2^31 / dimension
- Fixes ValueError: cannot reshape array for snapshots with ~2e9 particles
- Fix invalid escape sequences in docstrings by using raw strings

+++ pygad/gadget/handler.py
@@ -116,7 +116,7 @@ class FileReader(object):
         if self._format == 3:
             with h5py.File(self._filename, 'r') as gfile:
                 HDF5_name = config.std_name_to_HDF5.get(name,name)
-                data = np.empty(N*block.dimension, dtype=block.dtype)
+                data = np.empty(np.int64(N)*block.dimension, dtype=block.dtype)
                 if block.dimension > 1:
                     data = data.reshape( (N,block.dimension) )
                 off = 0
@@ -131,7 +131,7 @@ class FileReader(object):
                 gfile.seek(block.start_pos)
                 data = np.fromfile(gfile,
                                    dtype=self._endianness+block.type_descr,
-                                   count=N*block.dimension)
+                                   count=np.int64(N)*block.dimension)
                 if block.dimension > 1:
                     data = data.reshape( (N,block.dimension) )
